fix: move profile components to features/profile folder
