+++ src/features/profile/components/ProfileAvatarImageUrl.tsx
