+++ src/features/profile/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./ProfilePanel"
+export * from "./ProfileAvatarImageUrl"
+export * from "./ActionButtonSelector"
+export * from "./SetUpProfileModal"
